libs/os: Move into internal
