+++ abci/cmd/abci-cli/abci-cli.go
@@ -11,8 +11,8 @@ import (
 
 	"github.com/spf13/cobra"
 
+	cmtos "github.com/cometbft/cometbft/internal/os"
 	"github.com/cometbft/cometbft/libs/log"
-	cmtos "github.com/cometbft/cometbft/libs/os"
 
 	abcicli "github.com/cometbft/cometbft/abci/client"
 	"github.com/cometbft/cometbft/abci/example/kvstore"
