+++ cmd/cometbft/commands/rollback.go
@@ -9,9 +9,9 @@ import (
 	dbm "github.com/cometbft/cometbft-db"
 
 	cfg "github.com/cometbft/cometbft/config"
+	"github.com/cometbft/cometbft/internal/os"
 	"github.com/cometbft/cometbft/internal/state"
 	"github.com/cometbft/cometbft/internal/store"
-	"github.com/cometbft/cometbft/libs/os"
 )
 
 var removeBlock = false
