+++ cmd/priv_val_server/main.go
@@ -6,9 +6,9 @@ import (
 	"time"
 
 	"github.com/cometbft/cometbft/crypto/ed25519"
+	cmtos "github.com/cometbft/cometbft/internal/os"
 	"github.com/cometbft/cometbft/libs/log"
 	cmtnet "github.com/cometbft/cometbft/libs/net"
-	cmtos "github.com/cometbft/cometbft/libs/os"
 
 	"github.com/cometbft/cometbft/privval"
 )
