+++ proxy/multi_app_conn.go
@@ -4,8 +4,8 @@ import (
 	"fmt"
 
 	abcicli "github.com/cometbft/cometbft/abci/client"
+	cmtos "github.com/cometbft/cometbft/internal/os"
 	cmtlog "github.com/cometbft/cometbft/libs/log"
-	cmtos "github.com/cometbft/cometbft/libs/os"
 	"github.com/cometbft/cometbft/libs/service"
 )
